Modified base test case to remove all test data files.
Previously, if tests failed, randomly named test databases could
be left behind. This change makes it so that all test data files
are created within a directory which is created and removed after
tests have finished. In the case on an exception, a directory is
easier to manually clean up than many data files.

+++ chatterbot/adapters/storage/jsondatabase.py
@@ -123,5 +123,6 @@ class JsonDatabaseAdapter(StorageAdapter):
         """
         import os
 
-        os.remove(self.database.path)
+        if os.path.exists(self.database.path):
+            os.remove(self.database.path)
